move some types to frontend
